Reorganized FanCI Legacy Class
Reorganizing FanCI class

Reorganizing FanCI Class

Solving mask assignment bug

Added current_memory import
